Fix compiler warnings in test code
* GTest macro causes signed/unsigned integer comparison warnings.

Reviewed-on: http://review.couchbase.org/76002
Tested-by: Build Bot <build@couchbase.com>

+++ tests/gtest_internal_tests.cc
@@ -145,7 +145,8 @@ TEST_F(CouchstoreInternalTest, buffered_io_options)
                                                 0,
                                                 &Documents::docIterCheckCallback,
                                                 &documents));
-            EXPECT_EQ(docsInTest, documents.getCallbacks());
+            EXPECT_EQ(static_cast<int>(docsInTest),
+                      documents.getCallbacks());
             EXPECT_EQ(0, documents.getDeleted());
         }
